requirements.txt
arxiv
fastapi
uvicorn
weaviate-client[agents]
python-dotenv
# part 1
PyPDF2

# part 2
docling